havoc int to long

+++ trunk/source/Library-ModelCheckerAssistant/src/tw/ntu/svvrl/ultimate/lib/modelcheckerassistant/transitiontoolkit/threadtransitiontoolkit/ThreadStatementsExecutor.java
@@ -243,7 +243,7 @@ public class ThreadStatementsExecutor extends StatementsExecutor<ThreadState> {
 						updateThreadState(procName, varName, value);
 						break;
 					case BoogiePrimitiveType.INT:
-						value = r.nextInt();
+						value = r.nextLong();
 						updateThreadState(procName, varName, value);
 						break;
 					case BoogiePrimitiveType.REAL:
